Initial import
Just based off the jqr boilerplate for now.  C++ interface is going
to be a bit more challenging unfortunately, because there seems to be
an issue with including the headers (which use C99) - I think this
can be fine, but R is super strict about this.  I'll chase it up
later.